Moved APN_db.h under new cellular.
